Add first-class fit CLI command and wire existing flows
