+++ docs/cli.md
@@ -81,6 +81,7 @@ These remain fully supported for existing automation and muscle memory:
 - `gate`, `doctor`, `security`, `repo`, `evidence`, `report`, `policy`
 - `start` (guided alias to `onboarding` for role/journey-based first-run help)
 - `adoption` (follow-up action planner from fit + gate decision artifacts)
+- `fit` (risk-based fit recommendation for adoption depth)
 
 ### Supporting utilities (secondary)
